Working react-expo environment
I finally got this to export to my phone. Make sure that when you open the project you cd into the SeniorProject directory. We could probably change it so all the files in the SeniorProject are in the main directory and work it that way but I am too lazy atm. Once in that directory run npx expo start and that should start the server and you can scan the qr code to get it on your phone once you have the app.